# apps/server/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/manage/addresses/list": {
      "get": {
        "operationId": "AddressController_list",
        "parameters": [
          {
            "name": "province",
            "required": false,
            "in": "query",
            "description": "省份",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "city",
            "required": false,
            "in": "query",
            "description": "城市",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "district",
            "required": false,
            "in": "query",
            "description": "区/县",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "detail",
            "required": false,
            "in": "query",
            "description": "详细地址",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isDefault",
            "required": false,
            "in": "query",
            "description": "是否默认地址",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "consumerId",
            "required": false,
            "in": "query",
            "description": "客户ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取地址列表"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "获取地址列表",
        "tags": [
          "地址管理"
        ]
      }
    },
    "/api/v1/manage/addresses/page": {
      "get": {
        "operationId": "AddressController_page",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "页码",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "minimum": 1,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "province",
            "required": false,
            "in": "query",
            "description": "省份",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "city",
            "required": false,
            "in": "query",
            "description": "城市",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "district",
            "required": false,
            "in": "query",
            "description": "区/县",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "detail",
            "required": false,
            "in": "query",
            "description": "详细地址",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isDefault",
            "required": false,
            "in": "query",
            "description": "是否默认地址",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "consumerId",
            "required": false,
            "in": "query",
            "description": "客户ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取地址分页数据"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "分页获取地址列表",
        "tags": [
          "地址管理"
        ]
      }
    },
    "/api/v1/manage/addresses/detail/{id}": {
      "get": {
        "operationId": "AddressController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取地址详情"
          },
          "404": {
            "description": "地址不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "获取地址详情",
        "tags": [
          "地址管理"
        ]
      }
    },
    "/api/v1/manage/addresses/create": {
      "post": {
        "operationId": "AddressController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAddressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "地址创建成功"
          },
          "400": {
            "description": "请求参数错误"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "创建地址",
        "tags": [
          "地址管理"
        ]
      }
    },
    "/api/v1/manage/addresses/update": {
      "post": {
        "operationId": "AddressController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAddressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "地址更新成功"
          },
          "400": {
            "description": "请求参数错误"
          },
          "404": {
            "description": "地址不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "更新地址",
        "tags": [
          "地址管理"
        ]
      }
    },
    "/api/v1/manage/addresses/delete": {
      "post": {
        "operationId": "AddressController_delete",
        "parameters": [],
        "responses": {
          "200": {
            "description": "地址删除成功"
          },
          "400": {
            "description": "删除地址失败"
          },
          "404": {
            "description": "地址不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "删除地址",
        "tags": [
          "地址管理"
        ]
      }
    },
...
    "/api/v1/manage/commodity/list": {
      "get": {
        "operationId": "CommodityController_findAll",
        "parameters": [
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "商品名称",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "description",
            "required": false,
            "in": "query",
            "description": "商品描述",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "description": "最低价格",
            "schema": {
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "description": "最高价格",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "商品分类 ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "是否激活",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取商品列表"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "获取商品列表",
        "tags": [
          "商品管理"
        ]
      }
    },
    "/api/v1/manage/commodity/page": {
      "get": {
        "operationId": "CommodityController_findPage",
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功获取商品分页数据"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "分页获取商品列表",
        "tags": [
          "商品管理"
        ]
      }
    },
    "/api/v1/manage/commodity/detail/{id}": {
      "get": {
        "operationId": "CommodityController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取商品详情"
          },
          "404": {
            "description": "商品不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "获取商品详情",
        "tags": [
          "商品管理"
        ]
      }
    },
    "/api/v1/manage/commodity/create": {
      "post": {
        "operationId": "CommodityController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommodityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "商品创建成功"
          },
          "400": {
            "description": "请求参数错误"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "创建商品",
        "tags": [
          "商品管理"
        ]
      }
    },
    "/api/v1/manage/commodity/update": {
      "post": {
        "operationId": "CommodityController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommodityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "商品更新成功"
          },
          "400": {
            "description": "请求参数错误"
          },
          "404": {
            "description": "商品不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "更新商品",
        "tags": [
          "商品管理"
        ]
      }
    },
    "/api/v1/manage/commodity/delete": {
      "post": {
        "operationId": "CommodityController_delete",
        "parameters": [],
        "responses": {
          "200": {
            "description": "商品删除成功"
          },
          "404": {
            "description": "商品不存在"
          },
          "500": {
            "description": "删除商品失败"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "删除商品",
        "tags": [
          "商品管理"
        ]
      }
    },
    "/api/v1/manage/cart-items/list": {
      "get": {
        "operationId": "CartItemController_list",
        "parameters": [
          {
            "name": "consumerId",
            "required": false,
            "in": "query",
            "description": "客户ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commodityId",
            "required": false,
            "in": "query",
            "description": "商品ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取购物车列表"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "获取购物车列表",
        "tags": [
          "购物车管理"
        ]
      }
    },
    "/api/v1/manage/cart-items/page": {
      "get": {
        "operationId": "CartItemController_page",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "页码",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "minimum": 1,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "consumerId",
            "required": false,
            "in": "query",
            "description": "客户ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commodityId",
            "required": false,
            "in": "query",
            "description": "商品ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取购物车分页数据"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "分页获取购物车列表",
        "tags": [
          "购物车管理"
        ]
      }
    },
    "/api/v1/manage/cart-items/detail/{id}": {
      "get": {
        "operationId": "CartItemController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取购物车项详情"
          },
          "404": {
            "description": "购物车项不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "获取购物车项详情",
        "tags": [
          "购物车管理"
        ]
      }
    },
    "/api/v1/manage/cart-items/create": {
      "post": {
        "operationId": "CartItemController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCartItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "购物车项创建成功"
          },
          "400": {
            "description": "请求参数错误"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "创建购物车项",
        "tags": [
          "购物车管理"
        ]
      }
    },
    "/api/v1/manage/cart-items/update": {
      "post": {
        "operationId": "CartItemController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCartItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "购物车项更新成功"
          },
          "400": {
            "description": "请求参数错误"
          },
          "404": {
            "description": "购物车项不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "更新购物车项",
        "tags": [
          "购物车管理"
        ]
      }
    },
    "/api/v1/manage/cart-items/delete": {
      "post": {
        "operationId": "CartItemController_delete",
        "parameters": [],
        "responses": {
          "200": {
            "description": "购物车项删除成功"
          },
          "400": {
            "description": "删除购物车项失败"
          },
          "404": {
            "description": "购物车项不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "删除购物车项",
        "tags": [
          "购物车管理"
        ]
      }
    },
...
  },
  "info": {
    "title": "E-Commerce API",
    "description": "API documentation for the e-commerce management system",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateAddressDto": {
        "type": "object",
        "properties": {
          "consumerId": {
            "type": "string",
            "description": "客户ID"
          },
          "province": {
            "type": "string",
            "description": "省份"
          },
          "city": {
            "type": "string",
            "description": "城市"
          },
          "district": {
            "type": "string",
            "description": "区/县"
          },
          "detail": {
            "type": "string",
            "description": "详细地址"
          },
          "isDefault": {
            "type": "boolean",
            "description": "是否默认地址",
            "default": false
          }
        },
        "required": [
          "consumerId",
          "province",
          "city",
          "district",
          "detail"
        ]
      },
      "UpdateAddressDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "地址ID"
          }
        },
        "required": [
          "id"
        ]
      },
...
      "CommodityImageDto": {
        "type": "object",
        "properties": {
          "image": {
            "type": "string",
            "description": "商品图片URL"
          },
          "sortOrder": {
            "type": "number",
            "description": "排序顺序",
            "default": 0
          },
          "alt": {
            "type": "string",
            "description": "图片替代文本"
          }
        },
        "required": [
          "image"
        ]
      },
      "CreateCommodityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "商品名称"
          },
          "description": {
            "type": "string",
            "description": "商品描述"
          },
          "price": {
            "type": "number",
            "description": "商品价格",
            "minimum": 0
          },
          "stock": {
            "type": "number",
            "description": "商品库存",
            "minimum": 0
          },
          "mainImage": {
            "type": "string",
            "description": "商品主图片URL"
          },
          "categoryId": {
            "type": "string",
            "description": "商品分类 ID"
          },
          "isActive": {
            "type": "boolean",
            "description": "是否激活",
            "default": true
          },
          "images": {
            "description": "商品图片列表",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommodityImageDto"
            }
          }
        },
        "required": [
          "name",
          "price",
          "stock",
          "categoryId"
        ]
      },
      "UpdateCommodityDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "商品ID"
          }
        },
        "required": [
          "id"
        ]
      },
      "CreateCartItemDto": {
        "type": "object",
        "properties": {
          "consumerId": {
            "type": "string",
            "description": "客户ID"
          },
          "commodityId": {
            "type": "string",
            "description": "商品ID"
          },
          "quantity": {
            "type": "number",
            "description": "商品数量",
            "minimum": 1
          }
        },
        "required": [
          "consumerId",
          "commodityId",
          "quantity"
        ]
      },
      "UpdateCartItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "购物车项ID"
          }
        },
        "required": [
          "id"
        ]
      },
...
    }
  }
}